+++ docs/sitemap-0.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="UTF-8"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1"><url><loc>https://jakkunight.github.io/free-scada/</loc></url><url><loc>https://jakkunight.github.io/free-scada/api/</loc></url><url><loc>https://jakkunight.github.io/free-scada/guides/get-started/</loc></url></urlset>
