+++ docs/sitemap-index.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="UTF-8"?><sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"><sitemap><loc>https://jakkunight.github.io/free-scada/sitemap-0.xml</loc></sitemap></sitemapindex>
